+++ index.html
@@ -8,7 +8,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <div id="react-modals"></div>
+    <div id="burgermodal"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
