+++ src/components/ingredient-details/nutrition-value/styles.module.css
@@ -1,0 +1,6 @@
+.block {
+    align-items: center;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+}
