+++ src/components/modal-overlay/styles.module.css
@@ -7,5 +7,6 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    backdrop-filter: blur(5px);
+    backdrop-filter:
+        blur(5px) contrast(0.6);
 }
