+++ src/components/modal/modal-header/modal-header.tsx
@@ -12,7 +12,7 @@ import styles from './styles.module.css'
 export const ModalHeader: FC<Props> = (props: Props) => {
     return (
         <div className={styles.header}>
-            <a className="text text_type_main-medium">{props.headerText}</a>
+            <a className="text text_type_main-large">{props.headerText}</a>
             <CloseIcon type="primary" onClick={props.closeModal} className={styles.headerclose} />
         </div>
     )
